@@ -3,6 +3,7 @@ using System.Buffers;
 using System.Runtime.InteropServices;
 using System.Runtime.CompilerServices;
 using Pipelines.Sockets.Unofficial.Internal;
+using System.Diagnostics;
 
 namespace Pipelines.Sockets.Unofficial.Arenas
 {
@@ -21,6 +22,9 @@ namespace Pipelines.Sockets.Unofficial.Arenas
         /// </summary>
         public virtual void Clear(IMemoryOwner<T> allocation, int length)
             => allocation.Memory.Span.Slice(0, length).Clear();
+
+        internal virtual bool IsPinned => false;
+        internal virtual bool IsUnmanaged => false;
     }
 
     /// <summary>
@@ -62,7 +66,7 @@ namespace Pipelines.Sockets.Unofficial.Arenas
                 {
                     var arr = _array;
                     _array = null;
-                    if (arr != null) _pool.Return(arr);
+                    if (arr is not null) _pool.Return(arr);
                 }
             }
 
@@ -74,7 +78,7 @@ namespace Pipelines.Sockets.Unofficial.Arenas
 
             protected override bool TryGetArray(out ArraySegment<T> segment)
             {
-                if (_array != null)
+                if (_array is not null)
                 {
                     segment = new ArraySegment<T>(_array);
                     return true;
@@ -88,8 +92,14 @@ namespace Pipelines.Sockets.Unofficial.Arenas
         }
     }
 
-    internal sealed class PinnedArrayPoolAllocator<T> : Allocator<T> where T : unmanaged
+    internal sealed class PinnedArrayPoolAllocator<T> : Allocator<T>
     {
+        // where T : unmanaged
+        // is intended - can't enforce due to a: convincing compiler, and
+        // b: runtime (AOT) limitations
+
+        internal override bool IsPinned => true;
+
         private readonly ArrayPool<T> _pool;
 
         /// <summary>
@@ -97,7 +107,11 @@ namespace Pipelines.Sockets.Unofficial.Arenas
         /// </summary>
         public static PinnedArrayPoolAllocator<T> Shared { get; } = new PinnedArrayPoolAllocator<T>();
 
-        public PinnedArrayPoolAllocator(ArrayPool<T> pool = null) => _pool = pool ?? ArrayPool<T>.Shared;
+        public PinnedArrayPoolAllocator(ArrayPool<T> pool = null)
+        {
+            Debug.Assert(PerTypeHelpers<T>.IsBlittable);
+            _pool = pool ?? ArrayPool<T>.Shared;
+        }
 
         public override IMemoryOwner<T> Allocate(int length)
             => new PinnedArray(_pool, _pool.Rent(length));
@@ -107,42 +121,42 @@ namespace Pipelines.Sockets.Unofficial.Arenas
             private T[] _array;
             private readonly ArrayPool<T> _pool;
             private GCHandle _pin;
-            private T* _ptr;
+            private void* _ptr;
             public PinnedArray(ArrayPool<T> pool, T[] array)
             {
                 _pool = pool;
                 _array = array;
                 Length = array.Length;
                 _pin = GCHandle.Alloc(array, GCHandleType.Pinned);
-                _ptr = (T*)_pin.AddrOfPinnedObject().ToPointer();
+                _ptr = _pin.AddrOfPinnedObject().ToPointer();
             }
 
             protected override void Dispose(bool disposing)
             {
-                if (_ptr != null)
+                if (_ptr is not null)
                 {
                     _ptr = null;
-                    try { _pin.Free(); } catch { } // best efforst
+                    try { _pin.Free(); } catch { } // best efforts
                     _pin = default;
                 }
                 if (disposing)
                 {
                     var arr = _array;
                     _array = null;
-                    if (arr != null) _pool.Return(arr);
+                    if (arr is not null) _pool.Return(arr);
                     GC.SuppressFinalize(this);
                 }
             }
 
 
             public int Length { get; }
-            public override Span<T> GetSpan() => new Span<T>(_ptr, Length);
+            public override Span<T> GetSpan() => new(_ptr, Length);
 
-            public override MemoryHandle Pin(int elementIndex = 0) => new MemoryHandle(_ptr + elementIndex);
+            public override MemoryHandle Pin(int elementIndex = 0) => new(Unsafe.Add<T>(_ptr, elementIndex));
 
             protected override bool TryGetArray(out ArraySegment<T> segment)
             {
-                if (_array != null)
+                if (_array is not null)
                 {
                     segment = new ArraySegment<T>(_array);
                     return true;
@@ -158,9 +172,11 @@ namespace Pipelines.Sockets.Unofficial.Arenas
 
             void* IPinnedMemoryOwner<T>.Origin => _ptr;
 
+#pragma warning disable IDE0079
 #pragma warning disable CA2015 // possible GC while span in play; self-inflicted!
             ~PinnedArray() => Dispose(false);
 #pragma warning restore CA2015 // possible GC while span in play; self-inflicted!
+#pragma warning restore IDE0079
 
             public void Dispose() => Dispose(true);
         }
@@ -169,9 +185,17 @@ namespace Pipelines.Sockets.Unofficial.Arenas
     /// <summary>
     /// An allocator that allocates unmanaged memory, releasing the memory back to the OS when done
     /// </summary>
-    public unsafe sealed class UnmanagedAllocator<T> : Allocator<T> where T : unmanaged
+    public unsafe sealed class UnmanagedAllocator<T> : Allocator<T>
     {
-        private UnmanagedAllocator() { }
+        // where T : unmanaged
+        // is intended - can't enforce due to a: convincing compiler, and
+        // b: runtime (AOT) limitations
+        internal override bool IsUnmanaged => true;
+
+        private UnmanagedAllocator()
+        {
+            Debug.Assert(PerTypeHelpers<T>.IsBlittable);
+        }
 
         /// <summary>
         /// The global instance of the unmanaged allocator
@@ -185,22 +209,24 @@ namespace Pipelines.Sockets.Unofficial.Arenas
 
         private sealed class OwnedPointer : MemoryManager<T>, IPinnedMemoryOwner<T>
         {
+#pragma warning disable IDE0079
 #pragma warning disable CA2015 // possible GC while span in play; self-inflicted!
             ~OwnedPointer() => Dispose(false);
 #pragma warning restore CA2015 // possible GC while span in play; self-inflicted!
+#pragma warning restore IDE0079
 
-            private T* _ptr;
+            private void* _ptr;
 
             public int Length { get; }
             void* IPinnedMemoryOwner<T>.Origin => _ptr;
 
             public OwnedPointer(int length)
-                => _ptr = (T*)Marshal.AllocHGlobal((Length = length) * sizeof(T)).ToPointer();
+                => _ptr = Marshal.AllocHGlobal((Length = length) * Unsafe.SizeOf<T>()).ToPointer();
 
-            public override Span<T> GetSpan() => new Span<T>(_ptr, Length);
+            public override Span<T> GetSpan() => new(_ptr, Length);
 
             public override MemoryHandle Pin(int elementIndex = 0)
-                => new MemoryHandle(_ptr + elementIndex);
+                => new(Unsafe.Add<T>(_ptr, elementIndex));
 
             public override void Unpin() { } // nothing to do
 
@@ -214,7 +240,7 @@ namespace Pipelines.Sockets.Unofficial.Arenas
             {
                 var ptr = _ptr;
                 _ptr = null;
-                if (ptr != null) Marshal.FreeHGlobal(new IntPtr(ptr));
+                if (ptr is not null) Marshal.FreeHGlobal(new IntPtr(ptr));
                 if (disposing) GC.SuppressFinalize(this);
             }
         }
