@@ -115,7 +115,7 @@ namespace Pipelines.Sockets.Unofficial.Tests
 
             var sb = new StringBuilder();
             sb.Append($"{start.TryGetOffset()}-{ros.End.TryGetOffset()}; counts: ");
-            while (node != null & len > 0)
+            while (node is not null & len > 0)
             {
                 sb.Append("[").Append(node.RunningIndex).Append(',').Append(node.RunningIndex + node.Length).Append("):").Append(node.RefCount).Append(' ');
                 len -= node.Length;
