@@ -10,42 +10,38 @@ namespace Pipelines.Sockets.Unofficial.Tests
     public class ArenaTests
     {
 #pragma warning disable CS0169, IDE0051, RCS1213 // unused fields
-        private struct TwoPositions<T>
+        private readonly struct TwoPositions<T>
         {
             private readonly SequencePosition start, end;
         }
-        private struct TwoPair<T>
+        private readonly struct TwoPair<T>
         {
             private readonly object a, b;
             private readonly int c, d;
         }
-        private struct Len32<T>
+        private readonly struct Len32<T>
         {
             private readonly object x;
             private readonly int offset, len;
         }
-        private struct Len64<T>
+        private readonly struct Len64<T>
         {
             private readonly long len;
             private readonly object x;
-            private readonly int offse;
+            private readonly int offset;
         }
 #pragma warning restore CS0169, IDE0051, RCS1213
 
         [Fact]
         public void AssertPossibleLayoutSizes()
         {
-            int readOnlySequenceSize = 24;
-#if NETFRAMEWORK
-            // System.Memory nuget package misses this optimization
-            // present in .NET Core https://github.com/dotnet/corefx/pull/35860
-            readOnlySequenceSize = 32;
-#endif
-
-            // this test is re proving out the layout of Sequence<T>
             if (IntPtr.Size == 8)
             {
-                Assert.Equal(readOnlySequenceSize, Unsafe.SizeOf<ReadOnlySequence<int>>());
+#if NETFRAMEWORK // uses two-positions
+                Assert.Equal(32, Unsafe.SizeOf<ReadOnlySequence<int>>());
+#else           // uses two-pair
+                Assert.Equal(24, Unsafe.SizeOf<ReadOnlySequence<int>>());
+#endif
                 Assert.Equal(24, Unsafe.SizeOf<Sequence<int>>());
 
                 Assert.Equal(32, Unsafe.SizeOf<TwoPositions<int>>());
